Extend helpers/normalize.ts with project-level helpers
Move dedupeProjectIds, normalizeOperation, normalizeClamp, normalizeTab,
normalizeMachineDefinitions, cloneProject, instantiateProjectTemplate,
clearProjectMemoryCaches, projectsEqual, isFirstFeatureValid,
syncFeatureTreeProject, and syncStockFromSourceFeature from
projectStore.ts into helpers/normalize.ts. normalizeProject stays in
projectStore to avoid a cycle with backdropSlice. All slices now import
these directly instead of receiving them via DI.
