+++ src/store/projectStore.ts
@@ -15,13 +15,6 @@
  */
 
 import { create } from 'zustand'
-import { copyBundledDefinitions } from '../engine/gcode/definitions'
-import { validateMachineDefinition } from '../engine/gcode/types'
-import type { MachineDefinition } from '../engine/gcode/types'
-import {
-  clearImportedModelCaches,
-} from '../engine/importedMesh'
-import { clearSTLTransformedGeometryCache } from '../engine/csg'
 import { uniqueName } from '../import'
 import {
   defaultOrigin,
@@ -35,18 +28,13 @@ import {
   circleProfile,
 } from '../types/project'
 import type {
-  Clamp,
   FeatureOperation,
-  FeatureTreeEntry,
-  Operation,
   Project,
   SketchProfile,
   SketchFeature,
   PersistedImportedMesh,
-  Tab,
 } from '../types/project'
 import type { OpenProfileEndpoint } from './types'
-import { convertLength } from '../utils/units'
 import {
   clonePoint,
   pointsEqual,
@@ -66,16 +54,24 @@ import {
   previewOffsetFeatures as previewOffsetFeaturesWithFactory,
   type DerivedFeatureGroup,
 } from './helpers/derivedFeatures'
-import { idNumericSuffix, nextUniqueGeneratedId, syncIdCounter } from './helpers/ids'
+import { nextUniqueGeneratedId, syncIdCounter } from './helpers/ids'
 import {
+  cloneProject,
+  dedupeProjectIds,
+  normalizeClamp,
   normalizeFeatureZRange,
+  normalizeMachineDefinitions,
+  normalizeOperation,
+  normalizeTab,
   normalizeTool,
+  projectsEqual,
+  syncFeatureTreeProject,
+  syncStockFromSourceFeature,
 } from './helpers/normalize'
 import {
   transformProfile,
 } from './helpers/transform'
-import { isImportedModelFeature, normalizeImportedModelStorage, pruneUnusedModelAssets } from './helpers/modelAssets'
-import { fallbackOperationTarget, defaultOperationForTarget, isOperationTargetValid } from './helpers/operationDefaults'
+import { normalizeImportedModelStorage, pruneUnusedModelAssets } from './helpers/modelAssets'
 import { createPendingAddSlice } from './slices/pendingAddSlice'
 import { createPendingActionsSlice } from './slices/pendingActionsSlice'
 import { createPendingCompletionSlice } from './slices/pendingCompletionSlice'
@@ -267,339 +263,6 @@ function clearStaleConstraints(features: SketchFeature[], movedIds: Set<string>)
   return anyChanged ? next : features
 }
 
-export function syncFeatureTreeProject(project: Project): Project {
-  const featureFolders = project.featureFolders ?? []
-  const folderIdSet = new Set(featureFolders.map((folder) => folder.id))
-  const features = project.features.map((feature) => (
-    feature.folderId && !folderIdSet.has(feature.folderId)
-      ? { ...feature, folderId: null }
-      : feature
-  ))
-
-  const featureMap = new Map(features.map((feature) => [feature.id, feature]))
-  const usedRootFeatures = new Set<string>()
-  const usedFolders = new Set<string>()
-  const normalizedTree: FeatureTreeEntry[] = []
-
-  for (const entry of project.featureTree ?? []) {
-    if (entry.type === 'folder') {
-      if (folderIdSet.has(entry.folderId) && !usedFolders.has(entry.folderId)) {
-        normalizedTree.push(entry)
-        usedFolders.add(entry.folderId)
-      }
-      continue
-    }
-
-    const feature = featureMap.get(entry.featureId)
-    if (!feature || feature.folderId !== null || usedRootFeatures.has(entry.featureId)) {
-      continue
-    }
-
-    normalizedTree.push(entry)
-    usedRootFeatures.add(entry.featureId)
-  }
-
-  for (const folder of featureFolders) {
-    if (!usedFolders.has(folder.id)) {
-      normalizedTree.push({ type: 'folder', folderId: folder.id })
-      usedFolders.add(folder.id)
-    }
-  }
-
-  for (const feature of features) {
-    if (feature.folderId === null && !usedRootFeatures.has(feature.id)) {
-      normalizedTree.push({ type: 'feature', featureId: feature.id })
-      usedRootFeatures.add(feature.id)
-    }
-  }
-
-  const orderedFeatures: SketchFeature[] = []
-  const pushedFeatureIds = new Set<string>()
-
-  for (const entry of normalizedTree) {
-    if (entry.type === 'folder') {
-      for (const feature of features) {
-        if (feature.folderId === entry.folderId && !pushedFeatureIds.has(feature.id)) {
-          orderedFeatures.push(feature)
-          pushedFeatureIds.add(feature.id)
-        }
-      }
-      continue
-    }
-
-    const feature = featureMap.get(entry.featureId)
-    if (feature && !pushedFeatureIds.has(feature.id)) {
-      orderedFeatures.push(feature)
-      pushedFeatureIds.add(feature.id)
-    }
-  }
-
-  for (const feature of features) {
-    if (!pushedFeatureIds.has(feature.id)) {
-      orderedFeatures.push({ ...feature, folderId: null })
-    }
-  }
-
-  return {
-    ...project,
-    features: orderedFeatures,
-    featureFolders,
-    featureTree: normalizedTree,
-  }
-}
-
-/**
- * When a feature that serves as the stock source is modified, sync the stock
- * profile and thickness to match. Returns the updated project, or the original
- * if the featureId does not match the stock source.
- */
-export function syncStockFromSourceFeature(project: Project, featureId: string): Project {
-  const stock = project.stock
-  if (!stock.sourceFeature || stock.sourceFeatureId !== featureId) {
-    return project
-  }
-
-  // Find the updated source feature (it may be in features temporarily during sketch edit)
-  const updatedFeature = project.features.find((f) => f.id === featureId)
-  if (updatedFeature) {
-    // Feature was temporarily restored for editing; update sourceFeature copy.
-    // Use the feature's profile directly — it's already in world coordinates.
-    const syncedStock = {
-      ...stock,
-      sourceFeature: updatedFeature,
-      profile: updatedFeature.sketch.profile,
-      thickness: typeof updatedFeature.z_top === 'number' ? updatedFeature.z_top : stock.thickness,
-    }
-    return {
-      ...project,
-      stock: syncedStock,
-    }
-  }
-
-  // Feature is not in features array — sync from stock.sourceFeature directly
-  const source = stock.sourceFeature
-  return {
-    ...project,
-    stock: {
-      ...stock,
-      profile: source.sketch.profile,
-      thickness: typeof source.z_top === 'number' ? source.z_top : stock.thickness,
-    },
-  }
-}
-
-function dedupeProjectIds(project: Project): Project {
-  let localCounter = [
-    ...project.features.map((feature) => idNumericSuffix(feature.id)),
-    ...project.tools.map((tool) => idNumericSuffix(tool.id)),
-    ...project.operations.map((operation) => idNumericSuffix(operation.id)),
-    ...project.tabs.map((tab) => idNumericSuffix(tab.id)),
-    ...project.clamps.map((clamp) => idNumericSuffix(clamp.id)),
-  ].reduce((max, value) => Math.max(max, value), 0) + 1
-
-  const nextLocalId = (prefix: string) => `${prefix}${String(localCounter++).padStart(4, '0')}`
-
-  const seenFeatureIds = new Set<string>()
-  const features = project.features.map((feature) => {
-    if (!seenFeatureIds.has(feature.id)) {
-      seenFeatureIds.add(feature.id)
-      return feature
-    }
-
-    const nextId = nextLocalId('f')
-    return {
-      ...feature,
-      id: nextId,
-    }
-  })
-
-  const seenToolIds = new Set<string>()
-  const tools = project.tools.map((tool) => {
-    if (!seenToolIds.has(tool.id)) {
-      seenToolIds.add(tool.id)
-      return tool
-    }
-
-    const nextId = nextLocalId('t')
-    return {
-      ...tool,
-      id: nextId,
-    }
-  })
-
-  const seenOperationIds = new Set<string>()
-  const operations = project.operations.map((operation) => {
-    if (!seenOperationIds.has(operation.id)) {
-      seenOperationIds.add(operation.id)
-      return {
-        ...operation,
-      }
-    }
-
-    const nextId = nextLocalId('op')
-    return {
-      ...operation,
-      id: nextId,
-    }
-  })
-
-  const seenClampIds = new Set<string>()
-  const clamps = project.clamps.map((clamp) => {
-    if (!seenClampIds.has(clamp.id)) {
-      seenClampIds.add(clamp.id)
-      return { ...clamp }
-    }
-
-    const nextId = nextLocalId('cl')
-    return {
-      ...clamp,
-      id: nextId,
-    }
-  })
-
-  const seenTabIds = new Set<string>()
-  const tabs = project.tabs.map((tab) => {
-    if (!seenTabIds.has(tab.id)) {
-      seenTabIds.add(tab.id)
-      return { ...tab }
-    }
-
-    const nextId = nextLocalId('tb')
-    return {
-      ...tab,
-      id: nextId,
-    }
-  })
-
-  return {
-    ...project,
-    features,
-    tools,
-    operations,
-    tabs,
-    clamps,
-  }
-}
-
-function normalizeOperation(operation: Operation, project: Project, index: number): Operation {
-  const fallbackTarget = fallbackOperationTarget(project, operation.kind)
-  const defaults = defaultOperationForTarget(project, operation.kind, 'rough', fallbackTarget, index)
-  const normalized = {
-    ...defaults,
-    ...operation,
-    description: operation.description ?? '',
-    machiningOrder: operation.machiningOrder ?? 'level_first',
-    waterlineAdaptiveRefinement: operation.waterlineAdaptiveRefinement ?? true,
-    waterlineMicroStepover: operation.waterlineMicroStepover ?? 0,
-    waterlineRefinementThreshold: operation.waterlineRefinementThreshold ?? 0,
-    waterlineMaxRingsPerBand: operation.waterlineMaxRingsPerBand ?? 0,
-    waterlineTipStepdown: operation.waterlineTipStepdown ?? 0,
-  }
-
-  if (!isOperationTargetValid(project, normalized.kind, normalized.target)) {
-    return {
-      ...normalized,
-      target: fallbackTarget,
-    }
-  }
-
-  return normalized
-}
-
-function normalizeClamp(clamp: Clamp, units: Project['meta']['units'], index: number): Clamp {
-  const defaultSize = convertLength(12, 'mm', units)
-  const defaultHeight = convertLength(8, 'mm', units)
-  return {
-    id: clamp.id || `cl${index + 1}`,
-    name: clamp.name || `Clamp ${index + 1}`,
-    type: clamp.type ?? 'step_clamp',
-    x: clamp.x ?? 0,
-    y: clamp.y ?? 0,
-    w: Math.max(clamp.w ?? defaultSize, convertLength(0.1, 'mm', units)),
-    h: Math.max(clamp.h ?? defaultSize, convertLength(0.1, 'mm', units)),
-    height: Math.max(clamp.height ?? defaultHeight, convertLength(0.1, 'mm', units)),
-    visible: clamp.visible ?? true,
-  }
-}
-
-function normalizeTab(tab: Tab, units: Project['meta']['units'], index: number): Tab {
-  const defaultSize = convertLength(6, 'mm', units)
-  const defaultBottom = 0
-  const defaultTop = convertLength(3, 'mm', units)
-  const zBottom = tab.z_bottom ?? defaultBottom
-  const zTop = tab.z_top ?? defaultTop
-  return {
-    id: tab.id || `tb${index + 1}`,
-    name: tab.name || `Tab ${index + 1}`,
-    x: tab.x ?? 0,
-    y: tab.y ?? 0,
-    w: Math.max(tab.w ?? defaultSize, convertLength(0.1, 'mm', units)),
-    h: Math.max(tab.h ?? defaultSize, convertLength(0.1, 'mm', units)),
-    z_top: Math.max(zTop, zBottom),
-    z_bottom: Math.min(zTop, zBottom),
-    visible: tab.visible ?? true,
-  }
-}
-
-function normalizeMachineDefinitions(project: Project): {
-  machineDefinitions: MachineDefinition[]
-  selectedMachineId: string | null
-} {
-  const legacyMeta = project.meta as Project['meta'] & {
-    machineId?: string | null
-    customMachineDefinition?: MachineDefinition | null
-  }
-
-  const rawDefinitions = Array.isArray(project.meta.machineDefinitions)
-    ? project.meta.machineDefinitions
-    : null
-
-  if (!rawDefinitions) {
-    const machineDefinitions = copyBundledDefinitions()
-    let selectedMachineId: string | null = legacyMeta.machineId ?? null
-
-    if (legacyMeta.customMachineDefinition) {
-      const customDefinition = validateMachineDefinition({
-        ...legacyMeta.customMachineDefinition,
-        builtin: false,
-      })
-      machineDefinitions.push(customDefinition)
-      selectedMachineId = customDefinition.id
-    }
-
-    return {
-      machineDefinitions,
-      selectedMachineId: machineDefinitions.some((definition) => definition.id === selectedMachineId)
-        ? selectedMachineId
-        : null,
-    }
-  }
-
-  const definitions: MachineDefinition[] = []
-  const seenIds = new Set<string>()
-  for (const rawDefinition of rawDefinitions) {
-    try {
-      const definition = validateMachineDefinition(rawDefinition)
-      if (seenIds.has(definition.id)) {
-        continue
-      }
-      seenIds.add(definition.id)
-      definitions.push(definition)
-    } catch {
-      continue
-    }
-  }
-
-  const selectedMachineId = project.meta.selectedMachineId ?? null
-
-  return {
-    machineDefinitions: definitions,
-    selectedMachineId: definitions.some((definition) => definition.id === selectedMachineId)
-      ? selectedMachineId
-      : null,
-  }
-}
-
 export function normalizeProject(project: Project): Project {
   const modelAssets: Record<string, PersistedImportedMesh> = { ...(project.modelAssets ?? {}) }
   // Migration: convert 4-arc circles to native circle segments
@@ -689,64 +352,6 @@ export function normalizeProject(project: Project): Project {
   return normalizedProject
 }
 
-export function cloneProject(project: Project): Project {
-  const cloned = structuredClone(project)
-  cloned.modelAssets = project.modelAssets
-  return cloned
-}
-
-function instantiateProjectTemplate(template?: Project, name?: string): Project {
-  const now = new Date().toISOString()
-
-  if (!template) {
-    return newProject(name)
-  }
-
-  const cloned = cloneProject(template)
-  return {
-    ...cloned,
-    meta: {
-      ...cloned.meta,
-      name: name?.trim() || 'Untitled',
-      created: now,
-      modified: now,
-    },
-    backdrop: null,
-    dimensions: {},
-    annotations: [],
-    modelAssets: {},
-    features: [],
-    featureFolders: [],
-    featureTree: [],
-    global_constraints: [],
-    tools: [],
-    operations: [],
-    tabs: [],
-    clamps: [],
-    ai_history: [],
-  }
-}
-
-export function clearProjectMemoryCaches(): void {
-  clearImportedModelCaches()
-  clearSTLTransformedGeometryCache()
-}
-
-export function projectsEqual(a: Project, b: Project): boolean {
-  return JSON.stringify(a) === JSON.stringify(b)
-}
-
-// ============================================================
-// Rule: the first 2.5D feature must be 'add'.
-// Imported STL model features are standalone 3D model targets and may be the
-// only feature in a project, so they are exempt from the base-solid rule.
-// ============================================================
-
-export function isFirstFeatureValid(features: SketchFeature[]): boolean {
-  const firstMachiningFeature = features.find((feature) => feature.operation !== 'region')
-  if (!firstMachiningFeature) return true
-  return firstMachiningFeature.operation === 'add' || isImportedModelFeature(firstMachiningFeature)
-}
 
 // ============================================================
 // Store implementation
@@ -874,13 +479,10 @@ export const useProjectStore = create<ProjectStore>((rawSet, get) => {
     transactionStart: null,
   },
   ...createSelectionSlice(set, get, {
-    cloneProject,
     normalizeProject,
   }),
   ...createPendingActionsSlice(set),
   ...createPendingCompletionSlice(set, get, {
-    cloneProject,
-    projectsEqual,
     clearStaleConstraints,
     propagateConstraintsOnTranslate: (features, offsets) =>
       propagateConstraintsOnTranslate(features, offsets, { transformProfile }),
@@ -894,68 +496,37 @@ export const useProjectStore = create<ProjectStore>((rawSet, get) => {
       })
     },
     previewOffsetFeatures,
-    syncFeatureTreeProject,
     createDerivedFeature,
   }),
-  ...createPendingAddSlice(set, get, {
-    cloneProject,
-    syncFeatureTreeProject,
-  }),
-  ...createDimensionsSlice(set, get, { cloneProject }),
+  ...createPendingAddSlice(set, get),
+  ...createDimensionsSlice(set, get),
   ...createDimensionToolSlice(set, get),
-  ...createToolsSlice(set, get, { cloneProject, projectsEqual }),
-  ...createClampsSlice(set, get, { cloneProject, projectsEqual }),
-  ...createTabsSlice(set, get, { cloneProject, projectsEqual }),
-  ...createBackdropSlice(set, get, { cloneProject, projectsEqual }),
-  ...createMachineDefsSlice(set, get, { cloneProject, projectsEqual }),
-  ...createOperationsSlice(set, get, {
-    cloneProject,
-    projectsEqual,
-    syncFeatureTreeProject,
-  }),
-  ...createImportMergeSlice(set, get, {
-    cloneProject,
-    syncFeatureTreeProject,
-  }),
+  ...createToolsSlice(set, get),
+  ...createClampsSlice(set, get),
+  ...createTabsSlice(set),
+  ...createBackdropSlice(set),
+  ...createMachineDefsSlice(set),
+  ...createOperationsSlice(set, get),
+  ...createImportMergeSlice(set, get),
   ...createFeatureSlice(set, get, {
-    cloneProject,
-    syncFeatureTreeProject,
-    projectsEqual,
     createDerivedFeature,
-    syncStockFromSourceFeature,
   }),
   ...createFeatureGeometrySlice(set, get, {
-    cloneProject,
-    projectsEqual,
-    syncFeatureTreeProject,
-    syncStockFromSourceFeature,
     joinOpenProfiles,
     inferFeatureKind,
     clearStaleConstraints,
     applyProfileBreak,
   }),
-  ...createConstraintsSlice(set, get, {
-    cloneProject,
-  }),
-  ...createTreeVisibilitySlice(set, get, { cloneProject, projectsEqual }),
+  ...createConstraintsSlice(set),
+  ...createTreeVisibilitySlice(set),
   ...createProjectLifecycleSlice(set, get, {
     rawSet,
-    cloneProject,
-    projectsEqual,
     normalizeProject,
-    instantiateProjectTemplate,
-    clearProjectMemoryCaches,
   }),
   ...createHistorySlice(set, get, {
-    cloneProject,
-    projectsEqual,
     normalizeProject,
   }),
-  ...createWorkpieceSlice(set, get, {
-    cloneProject,
-    projectsEqual,
-    syncFeatureTreeProject,
-  }),
+  ...createWorkpieceSlice(set),
 
   }
 })
